requirements.txt
flask==2.2.2
pandas==1.5.3
jupyter==1.0.0
pip==22.3.1
panel==0.14.3
attrs==22.1.0
pathlib==1.0.1
wheel==0.38.4
cryptography==39.0.1
pillow==9.4.0
docutils==0.18.1
sphinx==5.0.2
tornado==6.1
jinja2==3.1.2
keyring==23.4.0
pytz==2022.7
typing_extensions==4.4.0
markupsafe==2.1.1
werkzeug==2.2.2
click==8.0.4
itsdangerous==2.0.1
zipp==3.11.0
mock==4.0.3
setuptools==65.6.3
packaging==22.0
pyparsing==3.0.9
future==0.18.3
six==1.16.0
requests==2.28.1
mysql-connector-python
Flask-Cors==3.0.10
PyJWT==2.2.0
gunicorn
